+++ examples/scaling/src/main.rs
@@ -237,14 +237,14 @@ fn spawn_bevy_ui(mut commands: Commands, asset_server: ResMut<AssetServer>) {
 fn spawn_screen_space(mut commands: Commands, asset_server: ResMut<AssetServer>) {
     commands
         .spawn((
-            VelloScreenSpace,
+            VelloRenderSpace::Screen,
             VelloScene::new(),
             Transform::from_xyz(CELL_WIDTH, SCREEN_HEIGHT / 2., 0.0),
             RotateThing,
         ))
         .with_children(|parent| {
             parent.spawn((
-                VelloScreenSpace,
+                VelloRenderSpace::Screen,
                 VelloTextSection {
                     value: "Scene in screen space".to_string(),
                     text_align: VelloTextAlign::Middle,
@@ -261,14 +261,14 @@ fn spawn_screen_space(mut commands: Commands, asset_server: ResMut<AssetServer>)
 
     commands
         .spawn((
-            VelloScreenSpace,
+            VelloRenderSpace::Screen,
             VelloSvgHandle(asset_server.load("embedded://scaling/assets/svg/fountain.svg")),
             Transform::from_xyz(CELL_WIDTH * 2., SCREEN_HEIGHT / 2., 0.0),
             RotateThing,
         ))
         .with_children(|parent| {
             parent.spawn((
-                VelloScreenSpace,
+                VelloRenderSpace::Screen,
                 VelloTextSection {
                     value: "SVG in screen space".to_string(),
                     text_align: VelloTextAlign::Middle,
@@ -285,7 +285,7 @@ fn spawn_screen_space(mut commands: Commands, asset_server: ResMut<AssetServer>)
 
     commands
         .spawn((
-            VelloScreenSpace,
+            VelloRenderSpace::Screen,
             VelloLottieHandle(asset_server.load("embedded://scaling/assets/lottie/Tiger.json")),
             Transform::from_xyz(CELL_WIDTH * 3., SCREEN_HEIGHT / 2., 0.0)
                 .with_scale(Vec3::splat(0.1)),
@@ -293,7 +293,7 @@ fn spawn_screen_space(mut commands: Commands, asset_server: ResMut<AssetServer>)
         ))
         .with_children(|parent| {
             parent.spawn((
-                VelloScreenSpace,
+                VelloRenderSpace::Screen,
                 VelloTextSection {
                     value: "Lottie in screen space".to_string(),
                     text_align: VelloTextAlign::Middle,
@@ -436,16 +436,16 @@ fn rotate(mut rotate_q: Query<&mut Transform, With<RotateThing>>, time: Res<Time
 
 #[allow(clippy::type_complexity)]
 fn gizmos(
-    svg: Query<(&VelloSvgHandle, &GlobalTransform), (Without<Node>, Without<VelloScreenSpace>)>,
-    lottie: Query<
-        (&VelloLottieHandle, &GlobalTransform),
-        (Without<Node>, Without<VelloScreenSpace>),
-    >,
+    svg: Query<(&VelloSvgHandle, &GlobalTransform, &VelloRenderSpace), Without<Node>>,
+    lottie: Query<(&VelloLottieHandle, &GlobalTransform, &VelloRenderSpace), Without<Node>>,
     svg_assets: Res<Assets<VelloSvg>>,
     lottie_assets: Res<Assets<VelloLottie>>,
     mut gizmos: Gizmos,
 ) {
-    for (svg, gtransform) in svg.iter() {
+    for (svg, gtransform, render_space) in svg.iter() {
+        if *render_space == VelloRenderSpace::Screen {
+            continue;
+        }
         let Some(svg) = svg_assets.get(svg.id()) else {
             continue;
         };
@@ -460,7 +460,10 @@ fn gizmos(
         );
     }
 
-    for (lottie, gtransform) in lottie.iter() {
+    for (lottie, gtransform, render_space) in lottie.iter() {
+        if *render_space == VelloRenderSpace::Screen {
+            continue;
+        }
         let Some(svg) = lottie_assets.get(lottie.id()) else {
             continue;
         };